term_setup: always restore initial stty settings on exit
term_setup/term_reset are used not only on start and on exit, but also
when switching from selection mode to search mode and back.

This has the unintended side effect of clobbering the stty variable with
the settings of selection mode when search mode is entered and thus
shellect returns to shell with -echo.

Fix this by defining two wrappers term_init and term_exit that are
exclusively called on startup and teardown respectively and are
unaffected by term_setup being called in-between.