LICENSE: relicense as GPL-2.0-or-later
To simplify distribution in projects that are GPL-2.0-only,
relicense the code to be GPL-2.0-or-later.

The main author of the code is Hui-Jun Chen with one outside
contribution from Ahmad Fatoum and both agree to relicense their
respective work.

The code also contains a snippet from shfm under MIT, which is
compatible. Replicate the copyright statement and link the original
repository.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>